backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
supabase==2.11.0
httpx==0.28.1
apscheduler==3.10.4
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5
google-api-python-client==2.159.0
google-auth==2.37.0
resend==2.5.1
python-dotenv==1.0.1
reportlab==4.0.9
